--- unbrake/src/actions/index.js
@@ -1,5 +1,0 @@
-export const add = (n1, n2) => ({
-	type: 'ADD', //Tipo da ação. Eu que escolho
-	n1,
-	n2
-})
