--- unbrake/src/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux'
-import sum from './sum'
-
-const Reducer = combineReducers({ //Agrupa os componentes
-	sum,
-})
-
-export default Reducer;
